requirements.txt
pillow>=10.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
